--- a/zstd_ddict.c
+++ b/zstd_ddict.c
@@ -26,7 +26,7 @@
 #include "zstd_ddict.h"
 
 #if defined(ZSTD_LEGACY_SUPPORT) && (ZSTD_LEGACY_SUPPORT>=1)
-#  include "zstd_legacy.h"
+#include "zstd_legacy.h"
 #endif