A bit different this time as zstd added some asm code ([huf_decompress_amd64.S](https://github.com/DataDog/zstd/commit/7c9ad24057a272848ee17995316ca6f348d7310d#diff-23596f2dd56fc29e6838cdc32534f8a2cb2698643145d152e25eecbb3216de55))

New workflow:
- [x] Run below:
```
git clone https://github.com/facebook/zstd.git c
cd c
git checkout v1.5.2
cd ..
git clone git@github.com:DataDog/zstd.git go
find c/lib \( -name '*.h' -o -name '*.c' -o -name '*.S' \) -exec cp -v {} go/ \;
```

- [x] Run `python tools/flatten_imports.py`
- [x] Run `python tools/insert_libzstd_ifdefs.py`


This is the code as of https://github.com/facebook/zstd/releases/tag/v1.5.2 (https://github.com/facebook/zstd/commit/e47e674cd09583ff0503f0f6defd6d23d8b718d3)

Benchmarks will come in the next comment