@Viq111 2022-04-15
The benchmarks on 
```
goos: darwin
goarch: amd64
pkg: github.com/DataDog/zstd
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
```
are interesting and not a slam dunk with `mr` payload.

```
goos: darwin
goarch: amd64
pkg: github.com/DataDog/zstd
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
name                    old time/op    new time/op    delta
BulkCompress-12           1.39µs ± 1%    1.47µs ± 1%   +5.81%  (p=0.000 n=18+20)
BulkDecompress-12          922ns ± 1%     921ns ± 1%     ~     (p=0.174 n=20+18)
CtxCompression-12         71.7ms ± 2%    77.4ms ± 1%   +7.83%  (p=0.000 n=17+18)
CtxDecompression-12       10.6ms ± 2%    10.3ms ± 1%   -2.41%  (p=0.000 n=18+20)
StreamCompression-12      80.2ms ± 1%    89.2ms ± 1%  +11.12%  (p=0.000 n=19+19)
StreamDecompression-12    12.8ms ± 1%    12.6ms ± 1%   -1.50%  (p=0.000 n=17+20)
Compression-12            72.0ms ± 1%    77.8ms ± 0%   +8.16%  (p=0.000 n=17+17)
Decompression-12          10.6ms ± 3%    10.3ms ± 1%   -3.01%  (p=0.000 n=19+19)

name                    old speed      new speed      delta
BulkCompress-12         66.3MB/s ± 1%  62.7MB/s ± 1%   -5.49%  (p=0.000 n=18+20)
BulkDecompress-12       61.8MB/s ± 1%  61.9MB/s ± 1%     ~     (p=0.169 n=20+18)
CtxCompression-12        139MB/s ± 2%   129MB/s ± 1%   -7.27%  (p=0.000 n=17+18)
CtxDecompression-12      942MB/s ± 2%   965MB/s ± 1%   +2.47%  (p=0.000 n=18+20)
StreamCompression-12     124MB/s ± 1%   112MB/s ± 1%  -10.01%  (p=0.000 n=19+19)
StreamDecompression-12   780MB/s ± 1%   792MB/s ± 1%   +1.52%  (p=0.000 n=17+20)
Compression-12           139MB/s ± 1%   128MB/s ± 0%   -7.55%  (p=0.000 n=17+17)
Decompression-12         937MB/s ± 3%   966MB/s ± 1%   +3.10%  (p=0.000 n=19+19)
```

So need to do benchmarks on other hardware and payload.