@Viq111 2022-04-22
Benchmark on AWS instances: `c5.2xlarge` for x86, `c6g.2xlarge` for ARM:

`mr` payload:
```
### AWS c5.2xl ###
name                   old time/op    new time/op    delta
BulkCompress-8           3.31µs ± 1%    3.43µs ± 1%   +3.70%  (p=0.000 n=20+20)
BulkDecompress-8         2.77µs ± 0%    2.78µs ± 1%   +0.52%  (p=0.001 n=20+20)
CtxCompression-8         89.3ms ± 1%    95.9ms ± 2%   +7.41%  (p=0.000 n=20+19)
CtxDecompression-8       12.9ms ± 1%    12.2ms ± 2%   -4.92%  (p=0.000 n=20+20)
StreamCompression-8       102ms ± 1%     113ms ± 2%  +10.47%  (p=0.000 n=20+20)
StreamDecompression-8    15.4ms ± 1%    14.9ms ± 2%   -3.19%  (p=0.000 n=20+20)
Compression-8            91.2ms ± 1%    95.5ms ± 1%   +4.75%  (p=0.000 n=19+20)
Decompression-8          12.8ms ± 1%    12.4ms ± 3%   -2.84%  (p=0.000 n=19+18)

name                   old speed      new speed      delta
BulkCompress-8         27.8MB/s ± 1%  26.8MB/s ± 1%   -3.57%  (p=0.000 n=20+20)
BulkDecompress-8       20.6MB/s ± 0%  20.5MB/s ± 1%   -0.51%  (p=0.001 n=20+20)
CtxCompression-8        112MB/s ± 1%   104MB/s ± 2%   -6.89%  (p=0.000 n=20+19)
CtxDecompression-8      775MB/s ± 1%   815MB/s ± 2%   +5.18%  (p=0.000 n=20+20)
StreamCompression-8    97.7MB/s ± 1%  88.5MB/s ± 2%   -9.47%  (p=0.000 n=20+20)
StreamDecompression-8   646MB/s ± 1%   667MB/s ± 2%   +3.30%  (p=0.000 n=20+20)
Compression-8           109MB/s ± 1%   104MB/s ± 1%   -4.54%  (p=0.000 n=19+20)
Decompression-8         781MB/s ± 1%   804MB/s ± 3%   +3.01%  (p=0.000 n=20+18)


### AWS c6g.2xl ###
name                   old time/op    new time/op    delta
BulkCompress-8           2.13µs ± 3%    2.18µs ± 1%  +2.79%  (p=0.000 n=20+20)
BulkDecompress-8         1.43µs ± 2%    1.43µs ± 2%    ~     (p=0.490 n=20+19)
CtxCompression-8          147ms ± 1%     151ms ± 2%  +2.79%  (p=0.000 n=20+20)
CtxDecompression-8       21.1ms ± 3%    20.3ms ± 2%  -4.03%  (p=0.000 n=20+20)
StreamCompression-8       165ms ± 3%     171ms ± 3%  +4.11%  (p=0.000 n=20+20)
StreamDecompression-8    25.0ms ± 2%    24.3ms ± 2%  -2.65%  (p=0.000 n=19+20)
Compression-8             145ms ± 3%     151ms ± 3%  +4.52%  (p=0.000 n=20+20)
Decompression-8          20.9ms ± 4%    20.1ms ± 2%  -3.90%  (p=0.000 n=20+20)

name                   old speed      new speed      delta
BulkCompress-8         43.3MB/s ± 3%  42.1MB/s ± 1%  -2.73%  (p=0.000 n=20+20)
BulkDecompress-8       39.9MB/s ± 2%  40.0MB/s ± 2%    ~     (p=0.465 n=20+19)
CtxCompression-8       67.8MB/s ± 1%  66.0MB/s ± 2%  -2.71%  (p=0.000 n=20+20)
CtxDecompression-8      472MB/s ± 3%   492MB/s ± 2%  +4.19%  (p=0.000 n=20+20)
StreamCompression-8    60.6MB/s ± 3%  58.2MB/s ± 3%  -3.96%  (p=0.000 n=20+20)
StreamDecompression-8   399MB/s ± 2%   410MB/s ± 2%  +2.73%  (p=0.000 n=19+20)
Compression-8          68.9MB/s ± 3%  65.9MB/s ± 3%  -4.32%  (p=0.000 n=20+20)
Decompression-8         476MB/s ± 3%   496MB/s ± 2%  +4.06%  (p=0.000 n=20+20)
```

Interpreted results:
- On x86, compression is 4-9% slower
- On x86, decompression is 0-5% faster
- On ARM, compression is 3-4% slower
- On ARM, decompression is 4% faster

Since we want to stay on upstream version, we should merge this PR but I will make a note of the performance impact in the release notes